Refactor: change _is_valid_config to no longer set the charm status
It was confusing to set the status of the charm due to configuration outside of the on_config_change event handler. This has now been improved and _is_valid_config not returns whether the condfiguration is valid and a message with a reason if it is not valid. Fully covered by tests.